////////////////////////////////////  ///                              ///  ///  USART PIN ASSIGNMENT        ///  ///                              ///  ///  Adafruit Metro M4 Express   ///  ////////////////////////////////////
Atmel Start project - basis of the entire program rests
on this functioning well with the current target board
(untested at this writing, but based on previous working
version - should require only setting 48 MHZ clock, and
maybe a few minor tweaks).

Has PINMUX specific to Adafruit Metro M4 Express, with
minimal pins assigned (TX, RX, LED (D13 location).

 30 #define PA16 GPIO(GPIO_PORTA, 16)
 31 #define PA22 GPIO(GPIO_PORTA, 22)
 32 #define PA23 GPIO(GPIO_PORTA, 23)
 33 #define PB14 GPIO(GPIO_PORTB, 14)

PA16 is the LED on the target, and is used to confirm basic
hardware function. If the clock isn't running, this won't
be able to blink (at all) as the MCU is effectively halted
(or something similar).

PA22 is D1 and Pin 43 of the ATSAMD51J19A and is assigned
to TX on SERCOM3 (Pad 0).

PA23 is D0 and Pin 44 and is assigned to RX on SERCOM3 (Pad 1).

 32 // UART on Metro M4 Express PA22 PA23
 33 // driver_init.c
 34 //     gpio_set_pin_function(PA22, PINMUX_PA22C_SERCOM3_PAD0); // TX
 35 //     gpio_set_pin_function(PA23, PINMUX_PA23C_SERCOM3_PAD1); // RX

TX is always assigned to PAD0 of any SERCOM on this chip.
RX may possibly be relocated? (onto PAD2) but TX cannot be
(new limitation specific to SAMD51, not present in SAMD21,
unless this is mistaken information).

On branch revision-00a-

	new file:   samd51_usart_08_sep_aa.atzip