atmel start files
These are demonstrations from the Atmel Start tool and in
this instance prints a 'Hello World' type of message to the
USART port.

This will be quickly replaced with its equivalent, to be
stored in src/cf/usart_gateway.c as part of the port of
CamelForth to the SAMD51 (and later, SAMD21 as well).

On branch revision-00a-

	new file:   driver_examples.c
	new file:   driver_examples.h