Replaces driver_init.c and driver_init.h
Use of '.inc' means the entire file is included into
another .inc or a .c source file, and so does not get
its own object file build in the Makefile.

This seems to be preferable to the author of CamelForth,
so it is done similarly, here.

On branch camelforth_port_aa

	new file:   usart_ini.h
	new file:   usart_ini.inc