On branch camelforth_port_bb
The full version of the locally authored Makefile.

It looks different, but it's fairly similar to the earlier one
it now replaces.

Meant to be streamlined for rapid iteration of

   make clean ; make ; make install

MANY OBJECT FILES are no longer built.  They will definitely
be needed for advanced hardware projects on this MCU.

They do not appear to be needed for the sync USART driver.

	modified:   Makefile