MASS DELETES cont
These deletes all seem to show a diff anyway.  Not sure
how that impacts memory when there's a lot of them.

In the past, gitk has slowed the computer to a crawl, by
trying to do too much in one go.

So, deletes in smaller groups, then.

On branch camelforth