Mass updates ItsyBitsy M4 Jan 2020
This directory is essentially the same as
the Trellis M4 directory.

Since ItsyBitsy M4 and NeoTrellis M4 are
just about the same, electrically, there
was no real need to modify.

The atmel start pins.h file was the whole deal,
and the only difference was that itsybitsym4 ignores
PA27 (Neopixel 3v on Trellis m4)

Don't believe this comment - believe the commits
prior to this one - already tested and runs well.

It may be possible therefore to use only what's been
committed so far, with the old (trellis m4) directory
and get a working program out of it.

But will mass update here as well (future commits).

January 1, 2020 02:00 UTC

	new file:   AtmelStart.env_conf
	new file:   AtmelStart.gpdsc
	new file:   README.md
	new file:   atmel_start.c
	new file:   atmel_start.h
	new file:   atmel_start_config.atstart
	new file:   driver_init.c
	new file:   driver_init.h
	new file:   main.c

On branch camelf_tada_tst_d_itsym4-a-